Remove password requirement (#117)
Some of our end users complained that the login process is too difficult. Remove the password requirement so it is easier to login. We can add it back later if we ever add features that are authenticated.

A problem this is that you can draw on another user's account (even when they're in the middle of a game). However, this was already an existing problem, because most users had no passwords anyway.